+++ home.html
@@ -45,8 +45,9 @@
         </Header>
         <main class="col-12 d-flex justify-content-center align-items-center flex-column">
             <div class="displayMSG"></div>
+            <h1>Home</h1>
             <div class="container new-post col-6 col-lg-5">
-                <h1>Create new post</h1>
+                <h2>Create new post</h2>
                 <form class="form-post postForm">
                     <div class="mb-4">
                         <label for="message-text" class="col-form-label">Post Title: <input name="title" id="title"
@@ -65,7 +66,7 @@
                 </form>
             </div>
             <div class="container list-post col-8 col-lg-5 justify-content-center align-items-center">
-                <h2>Post Feed:</h2>
+                <h3>Post Feed:</h3>
                 <div class="SeachBtnContainer">
                     <a href="search.html"> <button class="btn btn-outline-success" type="submit">Search for
                             posts</button></a>
