+++ js/home.mjs
@@ -2,15 +2,16 @@
 
 // variables
 import {
+  titleCreate,
   textCreate,
   createAndGetURL,
   postForm,
 } from "./modules/variables/variables.mjs";
 
 // fucntions
-import { success } from "./modules/posts/disableForm.mjs";
-import { fetchPosts } from "./modules/posts/getAllPosts.mjs";
-import { createPost } from "./modules/posts/createPosts.mjs";
+import { success } from "./modules/functions/posts/disableForm.mjs";
+import { fetchPosts } from "./modules/functions/posts/getAllPosts.mjs";
+import { createPost } from "./modules/functions/posts/createPosts.mjs";
 
 fetchPosts(createAndGetURL);
