+++ js/modules/functions/user/register.js
