+++ js/modules/functions/user/signout.js
