+++ update.html
@@ -47,6 +47,9 @@
             <div class="breadcrumb"><a href="home.html">Home</a> / <a class="currentID" href="#"></a></div>
 
             <div class="displayMSG"></div>
+            <div class="loader">
+                <p>Loading post, please wait...</p>
+            </div>
             <div class="container updateContainer new-post col-6 col-lg-5">
                 <h1 class="addPostId">Edit post: </h1>
                 <form class="form-post updateForm">
